Create index.html
The changes that was made there wasn't much to be changed in the process of creating this. The changes that I've made was to publish the initial commit by publishing the commit. This allows it to publish on the web. The second I was able to add the files that was requested of me. For the entire Index html there was no changes made to the file.